240420 이효진 유효성 검사 완료

+++ src/main/java/com/example/demo/materials/MaterialOutgoingVO.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class MaterialOutgoingVO {
 	private String prodInstructDetailCode;
 	private String matLotCode;
-	private Integer outCnt;
-	private Integer procNeedCnt;
+	private Double outCnt;
+	private Double procNeedCnt;
 	private String matCode;
 	private String result;
 }
